refactor and add option to specify position above/below option
