+++ src/components/PolystatPanel.tsx
@@ -108,6 +108,7 @@ export const PolystatPanel: React.FC<Props> = ({ options, data, id, width, heigh
           globalGradientsEnabled={options.globalGradientsEnabled}
           globalShowTimestampEnabled={options.globalShowTimestampEnabled}
           globalShowTimestampFormat={options.globalShowTimestampFormat}
+          globalShowTimestampPosition={options.globalShowTimestampPosition}
           globalShowTimestampYOffset={options.globalShowTimestampYOffset}
           globalTextFontAutoColor={options.globalTextFontAutoColorEnabled ? currentThemeV2.colors.text.primary : '#000000'}
           globalTextFontAutoColorEnabled={options.globalTextFontAutoColorEnabled}
