+++ src/components/alignment.test.ts
@@ -1,0 +1,13 @@
+import { GetAlignments } from './alignment';
+
+describe('Test GetAlignments', () => {
+  beforeEach(() => {
+
+  });
+
+  describe('Label Only', () => {
+    it('returns alignment when just label is showing', () => {
+    });
+  });
+
+});
