+++ src/components/auto_font_scaler.test.ts
@@ -1,0 +1,14 @@
+import { PolystatModel } from './types';
+import { AutoFontScalar } from './auto_font_scaler';
+
+describe('Test AutoFontScaler', () => {
+  beforeEach(() => {
+
+  });
+
+  describe('Scale for Label Only', () => {
+    it('returns font size for a labels', () => {
+    });
+  });
+
+});
